Refactor to use global methods and db folders
