+++ methods/client.js
@@ -1,7 +1,6 @@
 // import all methods here (export must be a function)
 import doSomething from './method1'
-
 // add to combined Meteor.methods, which will have the var name.
 Meteor.methods({
-  doSomething,
+  doSomething, 
 })
