+++ modules/db/Presentations.js
@@ -1,0 +1,3 @@
+const Presentations = new Mongo.Collection('presentations')
+
+export default Presentations
