--- modules/db/example.js
@@ -1,3 +1,0 @@
-const table = new Mongo.Collection('table')
-
-export default table
