+++ modules/main_Presenter/globals/example.js
@@ -1,10 +1,10 @@
-// const getSVG = Meteor.npmRequire('google-slides-downloader')
-import Presentations from 'main_SelectPresentation/globals/collections'
+// // const getSVG = Meteor.npmRequire('google-slides-downloader')
+// import Presentations from 'main_SelectPresentation/globals/collections'
 
 
-Meteor.methods({
-  // method for creating a new presentation in database with svg elements
-  changeIndex: function (gid, index, cb) {
-    Presentations.update({gid: gid.toString()}, {$set: {index: index}})
-  }
-})
+// Meteor.methods({
+//   // method for creating a new presentation in database with svg elements
+//   changeIndex: function (gid, index, cb) {
+//     Presentations.update({gid: gid.toString()}, {$set: {index: index}})
+//   }
+// })
